+++ CHANGES.md
@@ -3,8 +3,10 @@
 Unreleased
 
 - Use postponed evaluation of annotations ([pr #585][pr_585]).
+- Add static OpenAPI docs example ([issue #587][issue_587]).
 
 [pr_585]: https://github.com/apiflask/apiflask/pull/585
+[issue_587]: https://github.com/apiflask/apiflask/issues/587
 
 
 ## Version 2.1.3